Refactor code to move and rename cooldown skill effect (#266).
